Update README.md
In this Lab I will show you the process that I went through to responding to an alert on Splunk and my whole process of how I triage the event. The purpose of this lab is to show my understanding of proper security practices for responding to an incident. I will go over the specific event that was brought up on my alert on Splunk and my thought process on the analysis for the event.